Create Review Document
ABSTRACT
BASE PAPER ABSTRACT:
The objective of this paper is to apply machine learning and visual processing to identify overworked IT employees. Our technology is an improved version of older stress detection systems that did not include live detection or personal counseling. Stress detection methods that don't include real-time monitoring or individual counselling are being updated in this research. A survey is used to collect data on employees' mental stress levels in order to provide effective stress management solutions. In order to get the most out of your employees, this paper will look at stress management and how to create a healthy, spontaneous work environment.
OUR PROPOSED ABSTRACT:
In the modern world with latest technology gadgets, Stress is raising most to everyone. Because of this, despite their affluence, people are not satisfied. A pressured feeling is stress. Pressure may be mental, emotional, or even physical. Systems for managing stress are essential for identifying the stress levels that disturb our socioeconomic way of life. According to the World Health Organization (WHO), one in four people suffer from the mental health issue of stress. Human stress causes mental and socioeconomic issues, loss of focus at work, strained relationships with coworkers, despair, and in the worst circumstances, suicide. This requires the provision of counseling to help those under stress manage their stress. While it is impossible to completely avoid stress, taking preventive measures can help you manage it. Only medical and physiological professionals can now assess whether a person is depressed or stressed. A questionnaire-based approach is one of the more established ways to identify stress. Our project's primary goal is to identify signs of stress in IT professionals utilizing sophisticated machine learning and image processing methods. Our technology is an improved version of the previous stress detection technologies, which did not take into account the employee's emotions or live detection. However, this system includes both periodic and live employee emotion detection. Automatic detection of stress minimizes the risk of health issues and improves the welfare of the IT employee and the company. Knowing the IT employee's emotions allows the business to provide the right guidance and obtain better results from them. The accuracy of our suggested system model, which is developed using CNN Model Architecture, is 87.34% during training and 98.45% during validation.